+++ backend/src/api/endpoints/websocket_api.rs
@@ -267,6 +267,15 @@ async fn handle_event_message(socket: &mut WebSocket, event: EventMessage, handl
                             .await
                             .map_err(|e| format!("Playlist update progress event: {e} "))?;
                     }
+                    EventMessage::SystemInfoUpdate(system_info) => {
+                        let msg = ProtocolMessage::SystemInfoResponse(system_info)
+                            .to_bytes()
+                            .map_err(|e| e.to_string())?;
+                        socket
+                            .send(Message::Binary(msg))
+                            .await
+                            .map_err(|e| format!("System info event: {e} "))?;
+                    }
                 }
             }
         }