+++ backend/src/model/config/reverse_proxy.rs
@@ -8,6 +8,7 @@ use std::cmp::max;
 pub struct ReverseProxyDisabledHeaderConfig {
     pub referer_header: bool,
     pub x_header: bool,
+    pub cloudfare_header: bool,
     pub custom_header: Vec<String>,
 }
 
@@ -18,6 +19,9 @@ impl ReverseProxyDisabledHeaderConfig {
             return true;
         }
         if self.x_header && header_lc.starts_with("x-") {
+            return true;
+        }
+        if self.cloudfare_header && header_lc.starts_with("cf-") {
             return true;
         }
         self.custom_header
@@ -118,6 +122,7 @@ impl From<&ReverseProxyConfigDto> for ReverseProxyConfig {
             disabled_header: dto.disabled_header.as_ref().map(|d| ReverseProxyDisabledHeaderConfig {
                 referer_header: d.referer_header,
                 x_header: d.x_header,
+                cloudfare_header: d.cloudfare_header,
                 custom_header: d.custom_header.clone(),
             }),
             stream: dto.stream.as_ref().map(Into::into),
@@ -137,6 +142,7 @@ impl From<&ReverseProxyConfig> for ReverseProxyConfigDto {
             disabled_header: instance.disabled_header.as_ref().map(|d| ReverseProxyDisabledHeaderConfigDto {
                 referer_header: d.referer_header,
                 x_header: d.x_header,
+                cloudfare_header: d.cloudfare_header,
                 custom_header: d.custom_header.clone(),
             }),
             stream: instance.stream.as_ref().map(Into::into),