+++ shared/src/model/system_info.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq)]
+pub struct SystemInfo {
+    pub cpu_usage: f32,
+    pub memory_usage: u64,
+    pub memory_total: u64,
+}